+++ go/lib/fat/internal/filesystem_internal_test.go
@@ -51,11 +51,11 @@ func mockDirEntry(names mockDirEntryNames) Entry {
 
 type mockFatTable struct{}
 
-func (m mockFatTable) GetClusterTarget(_ uint16) uint16                    { panic("unused") }
-func (m mockFatTable) IsEoc(_ uint16) bool                                 { panic("unused") }
-func (m mockFatTable) GetMaxCluster() uint16                               { panic("unused") }
-func (m mockFatTable) GetEoc() uint16                                      { panic("unused") }
-func (m mockFatTable) WriteClusterTarget(_ *FileSystem, _, _ uint16) error { panic("unused") }
+func (m mockFatTable) GetClusterTarget(_ uint32) uint32                    { panic("unused") }
+func (m mockFatTable) IsEoc(_ uint32) bool                                 { panic("unused") }
+func (m mockFatTable) GetMaxCluster() uint32                               { panic("unused") }
+func (m mockFatTable) GetEoc() uint32                                      { panic("unused") }
+func (m mockFatTable) WriteClusterTarget(_ *FileSystem, _, _ uint32) error { panic("unused") }
 func mockParseFatTable(_ []byte) FatTable                                  { return mockFatTable{} }
 
 func mockGetRootDirectoryBytes(_ *FileSystem) ([]byte, error) { panic("unused") }
